Missed commiting the new filters.pl rule for sirc.
Essentially it's a set of rules that allows a user to easily
modify (via perl regex) messages.  Can use it to redirect it to
diffrent windows.

Hopefully I could use this to in the end to be able to completly undo
the hacks I've made to dsirc.... (hey you can always hope :)