3 set's of changes and update.
1. Biggest is the new KSTicker.  Window runs scroll from right to left
with the a set of lines. (Buffer is based on twice width).  Scroll
rate and speed are locked, this should be changed.

2. KSircMessageReceiver now has new function, set/getBroadcast.  If this
is set to TRUE they get broadcast traffic.  This is the default.  Set
false they do not.  Anything that does NOT need this traffic should
set it false to avoid broadcast storms.  of course ioBroadcaster is
set false. :)

3. DCC code should be cleaner and stabalized.