Adding it/ to Makefile.am and updating translations If only everything was is one place only...
Stephan, Eric Bischoff told me that you might plan to distribute the
translations in separate packages.
In this case, is it relevant to import translations to $app/doc, like I do now ?

I find the separate distribution a good idea, it would allow us to have
translations in only one place, www/documentation. But would there be
kdenetwork-fr, kdenetwork-de, ... ? That's a lot of packages...