Updated stable version to 2.3.6 (current non-CVS stable version).  This fixes tons o' bugs.. and adds some new features.  I've received a number of bug reports for the 2.11 version in 1.1.1 so a bug-fix was definitly needed. I had three options:
1) Ignore the bug reports and keep the 2.11 version to honor the freeze
2) Hack together yet another perversion of KBiff with *only* those bug
   fixes
3) Update to the newest stable version which contained all bug fixes

The first option was unacceptable.  I tried doing a test version of 2) but
was scared away from it for several reasons.  First, this version would not
have been tested at all (2.3.6 has been *extensively* tested) and second,
it would have resulted in Yet Another Code Split in the KBiff source.  I
am not willing to take either risk.

So it's all updated.  VERY unfortunately, this will break any number of
translations :-(  It is, IMO, a necessary evil in this case.